requirements.txt
fastapi~=0.110.2
pillow~=10.3.0
pytesseract~=0.3.10
python-multipart~=0.0.9
pylint~=3.1.0
uvicorn~=0.29.0
python-dotenv~=1.0.1
pytest~=8.1.1